chore: align Skill Store UI with PluginStore
