+++ src/locales/default/setting.ts
@@ -540,7 +540,8 @@ export default {
   'settingTTS.voice.preview': 'Voice Preview',
   'settingTTS.voice.title': 'Text-to-Speech Voice',
   'skillStore.button': 'Skill Store',
-  'skillStore.empty': 'No skills found',
+  'skillStore.empty': 'Browse the Skill store. Install one to get started, add more later.',
+  'skillStore.emptySearch': 'No matching Skills',
   'skillStore.networkError': 'Network error, please try again',
   'skillStore.search': 'Search skills...',
   'skillStore.tabs.community': 'Community',
